[#28] :recycle: refactor: 일기장 diary -> book 네이밍 변경
