+++ client/src/components/new-book/DiaryIcon.tsx
