+++ client/src/pages/NewBook.tsx
@@ -4,10 +4,10 @@ import { useRecoilValue } from "recoil";
 import { PurpleBackground } from "src/components/common/Backgrounds";
 import BackButton from "src/components/common/BackButton";
 import PageTitle from "src/components/common/PageTitle";
-import DiaryIcon from "src/components/new-diary/DiaryIcon";
-import ColorSelectContainer from "src/components/new-diary/ColorSelectContainer";
-import ContainerTitle from "src/components/new-diary/ContainerTitle";
-import DiaryTitleInputContainer from "src/components/new-diary/DiaryTitleInputContainer";
+import DiaryIcon from "src/components/new-book/DiaryIcon";
+import ColorSelectContainer from "src/components/new-book/ColorSelectContainer";
+import ContainerTitle from "src/components/new-book/ContainerTitle";
+import DiaryTitleInputContainer from "src/components/new-book/DiaryTitleInputContainer";
 import PurpleButton from "src/components/common/PurpleButton";
 import { selectedColorAtom, diaryTitleAtom } from "../recoil/new-diary";
 import { WARNING_MESSAGE } from "src/constants/WARNING_MESSAGE";
